refactor: extract GEXF export from the Apotheosis facade (closes #18)
draw(), add_attribute_schema() and save_gexf() move unchanged to a new
export::gexf module that consumes the facade through its public API.
