+++ src/controllers/apotheosis.rs
@@ -7,10 +7,9 @@ use crate::controllers::hnsw::Hnsw;
 use crate::controllers::radix_tree::RadixNode;
 use crate::datalayer::algorithms::DistanceAlgorithm;
 use crate::datalayer::record::{ApotheosisRecord, RadixKeyMapping};
-use gexf::{Edge, EdgeType, Gexf, Node as GefxNode};
-use std::fs::{self, File};
+use std::fs::File;
 use std::io::{Read, Write};
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 #[derive(serde::Serialize, serde::Deserialize)]
 #[serde(bound(
@@ -102,6 +101,21 @@ where
     /// Whether the model has no records indexed yet.
     pub fn is_empty(&self) -> bool {
         self.records.is_empty()
+    }
+
+    /// Structural view of the HNSW graph, one entry per layer: node feature
+    /// indices and their edges (source id, target id, distance). Needed by
+    /// `export::gexf`, which lives outside this impl and cannot reach `hnsw`
+    /// directly now that it is private.
+    #[allow(clippy::type_complexity)]
+    pub fn draw_model(&self) -> Vec<(Vec<usize>, Vec<(String, String, f32)>)> {
+        self.hnsw.draw_model()
+    }
+
+    /// Read-only access to a record by its index. Node indices returned by
+    /// `draw_model()` can be resolved to their full records through this.
+    pub fn record(&self, index: usize) -> Option<&R> {
+        self.records.get(index)
     }
 
     /// Performs an approximate k-NN search. If the `MetricId` natively maps to a Radix Tree key
@@ -199,92 +213,5 @@ where
 
         let decoded = bincode::deserialize_from(file)?;
         Ok(decoded)
-    }
-
-    // Exports the HNSW model to GEXF files for Gephi visualization.
-    /// Creates one file per layer with pattern `<path>_layer<N>.gexf`.
-    /// Nodes include all metadata attributes, edges represent HNSW connections.
-    ///
-    /// # Parameters
-    /// * `path` - Base filename for output (e.g., "model" creates "model_layer0.gexf", "model_layer1.gexf", etc.)
-    pub fn draw<P: AsRef<Path>>(&self, path: P) {
-        let base_path = path.as_ref();
-
-        let layer_gexfs = self.hnsw.draw_model();
-
-        // Enrich each layer with feature data and save
-        for (layer_idx, (nodes, edges)) in layer_gexfs.iter().enumerate() {
-            let mut gexf = Gexf::new(EdgeType::Undirected);
-            for node in nodes {
-                let mut gexf_node = GefxNode::new(node.to_string());
-                for (key, val) in self.records[*node].get_attributes() {
-                    gexf_node = gexf_node.with_attr(key, val);
-                }
-                let _ = gexf.add_node(gexf_node);
-            }
-
-            for (source_id, target_id, distance) in edges {
-                let _ = gexf.add_edge(
-                    Edge::new(
-                        format!("e_{}_{}", source_id, target_id),
-                        source_id.clone(),
-                        target_id.clone(),
-                    )
-                    .with_weight(*distance),
-                );
-            }
-
-            let _ = self.save_gexf(base_path, layer_idx, &gexf);
-        }
-    }
-
-    // Once draw is built, we need to add the attribute schema to the GEXF XML
-    // Has to be done manually since gexf crate does not support it yet.
-    fn add_attribute_schema(&self, xml: String, attributes: Vec<(String, String)>) -> String {
-        if let Some(graph_start) = xml.find("<graph")
-            && let Some(graph_end_offset) = xml[graph_start..].find(">")
-        {
-            let insert_pos = graph_start + graph_end_offset + 1;
-
-            let mut attrs = String::from("\n    <attributes class=\"node\">\n");
-            for (attribute_key, _) in attributes {
-                attrs.push_str(&format!(
-                    "      <attribute id=\"{}\" title=\"{}\" type=\"string\"/>\n",
-                    attribute_key, attribute_key
-                ));
-            }
-            attrs.push_str("    </attributes>\n");
-
-            let mut result = String::new();
-            result.push_str(&xml[..insert_pos]);
-            result.push_str(&attrs);
-            result.push_str(&xml[insert_pos..]);
-
-            return result;
-        }
-
-        xml
-    }
-
-    fn save_gexf(&self, base_path: &Path, layer_idx: usize, gexf: &Gexf) -> std::io::Result<()> {
-        let mut file_path = PathBuf::from(base_path);
-
-        if let Some(stem) = file_path.file_stem().and_then(|s| s.to_str()) {
-            let parent = file_path.parent().unwrap_or_else(|| Path::new(""));
-            let filename = format!("{}_layer{}.gexf", stem, layer_idx);
-            file_path = parent.join(filename);
-        } else {
-            let filename = format!("layer{}.gexf", layer_idx);
-            file_path = PathBuf::from(filename);
-        }
-
-        let fixed_xml = if !self.records.is_empty() {
-            self.add_attribute_schema(gexf.to_string().unwrap(), self.records[0].get_attributes())
-        } else {
-            gexf.to_string().unwrap()
-        };
-
-        fs::write(file_path, fixed_xml)?;
-        Ok(())
     }
 }
